refactor of routes + pages
