+++ plugs/index.js
@@ -1,10 +1,18 @@
 module.exports = {
   darkCrystal: {
     app: {
-      index: require('./app/index'),
-      show: require('./app/show')
+      page: {
+        index: require('./app/page/dark-crystal-index'),
+        show: require('./app/page/dark-crystal-show')
+      }
     },
-    router: require('./router'),
-    styles: require('./styles')
+    router: {
+      sync: {
+        routes: require('./router/sync/routes')
+      }
+    },
+    styles: {
+      mcss: require('./styles/mcss')
+    }
   }
 }
