+++ plugs/router/sync/routes.js
@@ -1,11 +1,10 @@
 const nest = require('depnest')
+const isRoot = require('scuttle-dark-crystal/isRoot')
 
 exports.gives = nest('router.sync.routes')
 exports.needs = nest({
-  'app.page': {
-    'darkCrystalIndex': 'first',
-    'darkCrystalShow': 'first'
-  }
+  'app.page.darkCrystalIndex': 'first',
+  'app.page.darkCrystalShow': 'first'
 })
 
 exports.create = (api) => {
@@ -14,7 +13,7 @@ exports.create = (api) => {
 
     const routes = [
       [ loc => loc.page === 'dark-crystal', pages.darkCrystalIndex ],
-      [ loc => loc.page === 'dark-crystal/show', pages.darkCrystalShow ]
+      [ loc => isRoot(loc), pages.darkCrystalShow ]
     ]
 
     return [...sofar, ...routes]
