+++ views/index.js
@@ -1,6 +1,5 @@
 const pull = require('pull-stream')
-const { h, Array: MutantArray, map, throttle, Value, when } = require('mutant')
-
+const { h, Array: MutantArray, map, throttle } = require('mutant')
 
 function DarkCrystalIndex (opts) {
   const {
@@ -15,10 +14,8 @@ function DarkCrystalIndex (opts) {
   ])
 
   function Root (msg) {
-    const show = Value(false)
-
     return h('div.crystal', [
-      h('div.overview', { 'ev-click': () => routeTo({ page: 'dark-crystal/show', root: msg }) }, [
+      h('div.overview', { 'ev-click': () => routeTo(msg) }, [
         h('div.name', msg.value.content.name),
         h('div.started', new Date(msg.value.timestamp).toLocaleDateString())
       ])
